Proposal for patching SonarQube rule 2184

+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri/src/org/eclipse/chemclipse/chromatogram/xxd/calculator/supplier/amdiscalri/impl/RetentionIndexCalculator.java
@@ -186,8 +186,8 @@ public class RetentionIndexCalculator {
 		 * Execute the calculation.
 		 */
 		float factorRetentionIndex = retentionIndexHigh - retentionIndexLow;
-		float nominatorRT = retentionTime - retentionTimeLow;
-		float denominatorRT = retentionTimeHigh - retentionTimeLow;
+		float nominatorRT = (float) retentionTime - retentionTimeLow;
+		float denominatorRT = (float) retentionTimeHigh - retentionTimeLow;
 		if(denominatorRT != 0) {
 			/*
 			 * Calculate the retention index.
